Make workdays-periodic-timer live up to its name
We have configured the maintainance windows for both our gke and eks
clusters to be all day Saturday. That's why it was very important to run
periodic tests only on workdays and Sunday. The workdays-periodic-timer
had the right name but not the right config, which lead to lots of test
failures on Saturday due to infrastrucure errors that were most likely
the result of maintainance.

Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>