Install latest golangci-lint in ci-image
We are currently pinned to golang 1.20.7, because go >=1.21.0 is making
our tests hang (both in github actions and concourse). Our linter job on
concourse broke down because it was being run on the official
golangci-lint image which was using golang v1.21.3. It is not a problem
to build golangci-lint with go >=1.21, but the `make lint` target is also
running `go vet ./...` which is failing unless we switch to go 1.21 in
the go.mod file, which we are not ready do do yet. In this commit we are
taking control of golangci-lint so that we can have latest linters
running on top of golang 1.20.7