Bump timeout for time consuming steps
The concourse timeout is a last resort to catch a potentially handing
job and get a ci notification. We are wrapping our e2e tests with a
timout of 30m already and sending them a SIGQUIT so that we see their
thread dumps. However concourse often kills our taks prematurely. We
suspect that concourse's timout counter starts ticking when the job is
scheduled meaning that if the job cannot start immediately (because
another job is running in the same serial group) it can get killed
prematurely.

Co-authored-by: Danail Branekov <danailster@gmail.com>