Retry cluster creation
We are doing this to avoid cloud related flakes like this one:
https://ci.korifi.cf-app.com/teams/main/pipelines/main/jobs/recreate-cluster-eks/builds/637#L664c59b9:1332

In essence we are creating a user and assingning the user some policy
via terraform, but sometimes there is a reace in which the arn of the
user does not become available for some time, which makes the policy
creation fail. Re-trying fixes this flake, since the user is created on
the first pass and the policy - on the socond pass and by that time the
arn is availlable.

This is likely an problem in AWS itself and not in Korifi or in our test
scripts, so re-trying is acceptable

Co-authored-by: Danail Branekov <danailster@gmail.com>