Mark build as successful after cleaning up the cluster
Sometimes a PR might intoduce a change that makes some resources leak,
for example a cfservicebinding with a stuck finalizer. This would cause
the undeploy-korifi step to fail, which is not recoverable without
manual intervention. If we mark the build as successful too early (as we
have been doing) such a change can sneak into main and cause CI problems
that require manual resolution.

Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>