Pin the ci-image Dockerfile to debian 12 bookworm
- This is done in order to fix the docker build, which fails on debian
  13 trixie as follows:

https://ci.korifi.cf-app.com/teams/main/pipelines/ci-images/jobs/update-ci-image/builds/557#L686255b9:93

- The reason for the failure is that ubuntu images based on trixie no
  longer iclude the apt-add binary which is deprecated in favour of
  using gpg
- We could simply switch to gpg, but the Dockerfile installs terraform
  from the hashicorp apt repository and as it turns out hashicorp have
  not published a "trixie" version of their deb repos
- The easiest thing to do is to pin ourselves to "bookworm" until thore
  repos are published
- Note that we are not pinning the golang version, only the debian
  version. Pinning to a particular debian version is actually
  recommended by the official golang docker repo:

Some of these [golang] tags may have names like bookworm or trixie in
them. These are the suite code names for releases of Debian and indicate
which release the image is based on. If your image needs to install any
additional packages beyond what comes with the image, you'll likely want
to specify one of these explicitly to minimize breakage when there are
new releases of Debian.

We might want to revert this once terraform is available for trixie