Change periodics schedule
This is a part of the cost cutting initiative

It turns out that we are paying the same even if we reduce the number of
nightly test runs by half. This means that we are baing charged more
based on uptime than real load of the clusters. So if we have about 8
test runs each  taking 30-40m, we can run them in a tigth sequence from
7pm until midnight, then destroy the test clusters and re-create them in
the morning. This means that we will save about 7h of uptime which is
about 30% of one day.