Recreate the `korifi-gateway` namespace when deploying local korifi
Now that the gateway namespace should be created by the user before
installing Korifi, we need to ensure that the namespace exists before
installing the local korifi version. We still need to annotate it with
`helm.sh/resource-policy: keep` in order for the upgrade from latest
release not to wipe it out.

When going back to the latest release we need to delete the namespace
before installing latest. Otherwise, helm would complain that the
namespace used during test run is not managed by helm

Related PR: https://github.com/cloudfoundry/korifi/pull/4214